+++ src/Type/Redactor.php
@@ -8,15 +8,5 @@ use Melting\MODX\RTE\BaseRTE;
 class Redactor extends BaseRTE
 {
     protected $override = 'redactor-overrides.js';
-
-    /**
-     * @inherit
-     */
-    public function getOptions()
-    {
-        $settings = $this->rte->getRTEOptions();
-        foreach ($settings as $k) {
-            $this->modx->setOption("redactor.{$k}", $this->getSetting($k));
-        }
-    }
+    protected $prefix = 'redactor';
 }