+++ src/Type/TinyMCERTE.php
@@ -8,19 +8,5 @@ use Melting\MODX\RTE\BaseRTE;
 class TinyMCERTE extends BaseRTE
 {
     protected $override = 'tinymcerte-overrides.js';
-
-    /**
-     * @inherit
-     */
-    public function getOptions()
-    {
-        $settings = $this->rte->getRTEOptions();
-        foreach ($settings as $k) {
-            $this->modx->setOption("tinymcerte.{$k}", $this->getSetting($k));
-        }
-
-        return [
-            'editor' => 'TinyMCE RTE',
-        ];
-    }
+    protected $prefix = 'tinymcerte';
 }