+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Changes for MODX RTE library
 
 ## [unreleased] - xx/xx/2016
 
+* Added the ability to define per field configuration
 * Wrapped the overrides in a function to make sure `MODx.loadRTE` (RTE implementations) is available
 * Allow Redactor RTE to have a fixed width
 * Allow Redactor RTE to have a fixed height (no auto grow)